check if entity exists before trying to delete by id

+++ src/main/java/com/uav_recon/app/api/services/InspectionService.kt
@@ -78,6 +78,9 @@ class InspectionService(
     }
 
     fun delete(id: String) {
+        if (!inspectionRepository.findById(id).isPresent) {
+            throw Error(101, "Invalid inspection UUID")
+        }
         inspectionRepository.deleteById(id)
     }
